backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clerk-backend-api>=1.6.0",
    "fastapi>=0.119.1",
    "openai>=2.2.0",
    "python-dotenv>=1.0.1",
    "ruff>=0.14.2",
    "sqlalchemy>=2.0.44",
    "uvicorn>=0.33.0",
]
